+++ examples/disconnects/client.py
@@ -1,6 +1,7 @@
 from __future__ import print_function
 import sys
 import Pyro4
+from Pyro4.core import MessageFactory
 
 if sys.version_info < (3, 0):
     input = raw_input
@@ -16,13 +17,16 @@ class AutoReconnectingProxy(Pyro4.core.Proxy):
     the server to see if it still has a working connection. If not, it
     reconnects the proxy and retries the method call.
     """
-    _pyroPingMethod = "ping"    # replace by daemon method?
 
     def _pyroInvoke(self, methodname, vargs, kwargs, flags=0):
         if self._pyroConnection:
             try:
                 print("  <proxy: ping>")
-                super(AutoReconnectingProxy, self)._pyroInvoke(self._pyroPingMethod, [], {})
+                # send the special 'ping' message to the daemon, expecting a 'ping' response (no-op)
+                data = MessageFactory.createMessage(MessageFactory.MSG_PING, None, 0, 0)
+                self._pyroConnection.send(data)
+                MessageFactory.getMessage(self._pyroConnection, [MessageFactory.MSG_PING])
+                print("  <proxy: ping reply (still connected)>")
             except Pyro4.errors.ConnectionClosedError:     # or possibly even ProtocolError
                 print("  <proxy: Connection lost. REBINDING...>")
                 self._pyroReconnect()